Factor metrics snapshotter out of http server
This will allow the same metrics serilization code to be used for the HTTP
endpoint and for writing to kafka and any other push modules that we write.
