+++ src/stats/mod.rs
@@ -13,6 +13,208 @@ use metrics::*;
 use std::fs::{File, OpenOptions};
 use std::io::Write;
 
+pub struct MetricsSnapshot {
+    metrics: Metrics<AtomicU32>,
+    snapshot: Vec<Reading>,
+    refreshed: u64,
+    count_label: Option<String>,
+}
+
+impl MetricsSnapshot {
+    pub fn new(
+        metrics: Metrics<AtomicU32>,
+        count_label: Option<&str>,
+    ) -> Self {
+        Self {
+            metrics,
+            snapshot: Vec::new(),
+            refreshed: 0,
+            count_label: count_label.map(std::string::ToString::to_string),
+        }
+    }
+
+    pub fn refresh(&mut self) {
+            self.snapshot = self.metrics.readings();
+            self.refreshed = time::precise_time_ns();
+    }
+
+    pub fn prometheus(&self) -> String {
+        let mut data = Vec::new();
+        for reading in &self.snapshot {
+            let label = reading.label();
+            let output = reading.output();
+            let value = reading.value();
+            match output {
+                Output::Counter => {
+                    if let Some(ref count_label) = self.count_label {
+                        data.push(format!("{}/{} {}", label, count_label, value));
+                    } else {
+                        data.push(format!("{} {}", label, value));
+                    }
+                }
+                Output::Percentile(percentile) => match percentile {
+                    Percentile::Minimum => {
+                        data.push(format!("{}/minimum/value {}", label, value));
+                    }
+                    Percentile::Maximum => {
+                        data.push(format!("{}/maximum/value {}", label, value));
+                    }
+                    _ => {
+                        data.push(format!("{}/histogram/{} {}", label, percentile, value));
+                    }
+                },
+                Output::MaxPointTime => {
+                    // we have point's ns since X and current timespec and current ns sinc X
+                    let point_ns = value;
+                    let now_timespec = time::get_time();
+                    let now_ns = time::precise_time_ns();
+
+                    // find the number of NS in the past for point
+                    let delta_ns = now_ns - point_ns;
+                    let point_timespec =
+                        now_timespec - time::Duration::nanoseconds(delta_ns as i64);
+
+                    // convert to UTC
+                    let point_utc = time::at_utc(point_timespec);
+                    // calculate offset from the top of the minute
+                    let offset = point_utc.tm_sec as u64 * 1_000_000_000 + point_utc.tm_nsec as u64;
+                    let offset_ms = (offset as f64 / 1_000_000.0).floor() as u64;
+                    data.push(format!("{}/maximum/offset_ms {}", label, offset_ms));
+                }
+                _ => {
+                    continue;
+                }
+            }
+        }
+        data.sort();
+        let mut content = data.join("\n");
+        content += "\n";
+        let parts: Vec<&str> = content.split('/').collect();
+        parts.join("_")
+    }
+
+    pub fn human(&self) -> String {
+        let mut data = Vec::new();
+        for reading in &self.snapshot {
+            let label = reading.label();
+            let output = reading.output();
+            let value = reading.value();
+            match output {
+                Output::Counter => {
+                    if let Some(ref count_label) = self.count_label {
+                        data.push(format!("{}/{}: {}", label, count_label, value));
+                    } else {
+                        data.push(format!("{}: {}", label, value));
+                    }
+                }
+                Output::Percentile(percentile) => match percentile {
+                    Percentile::Minimum => {
+                        data.push(format!("{}/minimum/value: {}", label, value));
+                    }
+                    Percentile::Maximum => {
+                        data.push(format!("{}/maximum/value: {}", label, value));
+                    }
+                    _ => {
+                        data.push(format!("{}/histogram/{}: {}", label, percentile, value));
+                    }
+                },
+                Output::MaxPointTime => {
+                    // we have point's ns since X and current timespec and current ns sinc X
+                    let point_ns = value;
+                    let now_timespec = time::get_time();
+                    let now_ns = time::precise_time_ns();
+
+                    // find the number of NS in the past for point
+                    let delta_ns = now_ns - point_ns;
+                    let point_timespec =
+                        now_timespec - time::Duration::nanoseconds(delta_ns as i64);
+
+                    // convert to UTC
+                    let point_utc = time::at_utc(point_timespec);
+                    // calculate offset from the top of the minute
+                    let offset = point_utc.tm_sec as u64 * 1_000_000_000 + point_utc.tm_nsec as u64;
+                    let offset_ms = (offset as f64 / 1_000_000.0).floor() as u64;
+                    data.push(format!("{}/maximum/offset_ms: {}", label, offset_ms));
+                }
+                _ => {
+                    continue;
+                }
+            }
+        }
+        data.sort();
+        let mut content = data.join("\n");
+        content += "\n";
+        content
+    }
+
+    fn json(&self, pretty: bool) -> String {
+        let mut head = "{".to_owned();
+        if pretty {
+            head += "\n  ";
+        }
+        let mut data = Vec::new();
+        for reading in &self.snapshot {
+            let label = reading.label();
+            let output = reading.output();
+            let value = reading.value();
+            match output {
+                Output::Counter => {
+                    if let Some(ref count_label) = self.count_label {
+                        data.push(format!("\"{}/{}\": {}", label, count_label, value));
+                    } else {
+                        data.push(format!("\"{}\": {}", label, value));
+                    }
+                }
+                Output::Percentile(percentile) => match percentile {
+                    Percentile::Minimum => {
+                        data.push(format!("\"{}/minimum/value\": {}", label, value));
+                    }
+                    Percentile::Maximum => {
+                        data.push(format!("\"{}/maximum/value\": {}", label, value));
+                    }
+                    _ => {
+                        data.push(format!("\"{}/histogram/{}\": {}", label, percentile, value));
+                    }
+                },
+                Output::MaxPointTime => {
+                    // we have point's ns since X and current timespec and current ns since X
+                    let point_ns = value;
+                    let now_timespec = time::get_time();
+                    let now_ns = time::precise_time_ns();
+
+                    // find the number of NS in the past for point
+                    let delta_ns = now_ns - point_ns;
+                    let point_timespec =
+                        now_timespec - time::Duration::nanoseconds(delta_ns as i64);
+
+                    // convert to UTC
+                    let point_utc = time::at_utc(point_timespec);
+                    // calculate offset from the top of the minute
+                    let offset = point_utc.tm_sec as u64 * 1_000_000_000 + point_utc.tm_nsec as u64;
+                    let offset_ms = (offset as f64 / 1_000_000.0).floor() as u64;
+                    data.push(format!("\"{}/maximum/offset_ms\": {}", label, offset_ms));
+                }
+                _ => {
+                    continue;
+                }
+            }
+        }
+        data.sort();
+        let body = if pretty {
+            data.join(",\n  ")
+        } else {
+            data.join(",")
+        };
+        let mut content = head;
+        content += &body;
+        if pretty {
+            content += "\n";
+        }
+        content += "}";
+        content
+    }
+}
+
 pub struct StatsLog {
     file: File,
     metrics: Metrics<AtomicU32>,
